pyproject.toml
[project]
name = "deployed-a2a-sample"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk[a2a]>=1.22.1",
    "rich>=13.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.2.1",
]
